+++ kernel-rs/src/memlayout.rs
@@ -17,7 +17,7 @@ use crate::riscv::{MAXVA, PGSIZE};
 /// PHYSTOP -- end RAM used by the kernel
 
 /// qemu puts UART registers here in physical memory.
-pub const UART0: u64 = 0x10000000;
+pub const UART0: usize = 0x10000000;
 pub const UART0_IRQ: i32 = 10;
 
 /// virtio mmio interface
@@ -26,8 +26,8 @@ pub const VIRTIO0_IRQ: i32 = 1;
 
 /// local interrupt controller, which contains the timer.
 pub const CLINT: i64 = 0x2000000;
-pub const fn clint_mtimecmp(hartid: u64) -> u64 {
-    (CLINT + 0x4000 as i64 + (8 * hartid) as i64) as u64
+pub const fn clint_mtimecmp(hartid: usize) -> usize {
+    (CLINT + 0x4000 as i64 + (8 * hartid) as i64) as usize
 }
 pub const CLINT_MTIME: i64 = CLINT + 0xbff8 as i32 as i64;
